+++ src/utils/StateStorage.ts
@@ -71,7 +71,9 @@ export const DEFAULT_EDIT_SETTING: EditSetting = {
   proofreadApiUrl: 'https://api.languagetool.org/v2/check',
   proofreadApiKey: '',
   workspaceIgnoreRules: DEFAULT_WORKSPACE_IGNORE_RULES,
-  useGitignoreAsWorkspaceIgnore: true,
+  useGitignoreForExplorer: false,
+  useGitignoreForSearch: false,
+  useGitignoreForWatcher: false,
   codeBlockLanguageScope: 'common',
 }